refactor: Discount 클래스의 잘못된 패키지 경로 수정
