+++ app/src/main/java/woowacourse/shopping/feature/order/order/OrderPresenter.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.feature.order.order
 
+import com.example.domain.model.Discount
 import com.example.domain.repository.CartRepository
 import com.example.domain.repository.OrderRepository
-import woowacourse.shopping.feature.order.Discount
 import woowacourse.shopping.mapper.toPresentation
 import woowacourse.shopping.model.CartProductUiModel
