+++ domain/src/main/java/com/example/domain/model/MerchandiseItem.kt
@@ -1,7 +1,5 @@
 package com.example.domain.model
 
-import woowacourse.shopping.feature.order.Discount
-
 interface MerchandiseItem {
     fun use(): Pair<Int, Discount.Condition?>
 }
